refactor : RecreatedBackground dto 구조 변경
