+++ TecheerPicture/src/main/java/com/techeerpicture/TecheerPicture/RecreatedBackground/entity/RecreatedBackground.java
@@ -1,4 +1,4 @@
-package com.techeerpicture.TecheerPicture.RecreatedBackground;
+package com.techeerpicture.TecheerPicture.RecreatedBackground.entity;
 
 import jakarta.persistence.*;
